refactor: remove old implementations
