+++ src/index.test.ts
@@ -1,0 +1,17 @@
+import {
+  useFn,
+  useInterceptor,
+  useWire,
+  useWireState,
+  useWireValue,
+} from './index';
+
+describe('index file', () => {
+  it('should exports correctly', () => {
+    expect(useWire).toBeDefined();
+    expect(useWireValue).toBeDefined();
+    expect(useWireState).toBeDefined();
+    expect(useInterceptor).toBeDefined();
+    expect(useFn).toBeDefined();
+  });
+});
