--- src/interceptor.ts
@@ -1,4 +1,0 @@
-export type Interceptor<Value> = (
-  nextValue: Value,
-  preValue: Value | undefined,
-) => Value | undefined;
