+++ src/types.d.ts
@@ -1,6 +1,4 @@
-export { Subscribable } from './subscribable';
-export { Wire } from './wire';
-export { Interceptor } from './interceptor';
-
-// TODO remove. kept for backward compatibility
-export { Interceptor as Interferer } from './interceptor';
+export { StateWire, WireState } from './state-wire/state-wire';
+export { FnsWire, WireFns } from './fn-wire/fns-wire';
+export { Wire } from './wire/wire';
+export { Interceptor } from './interceptor/interceptor';
