--- src/utils/batched-updates.ts
@@ -1,5 +1,0 @@
-import ReactDOM from 'react-dom';
-
-export function batchedUpdates(fn: () => void) {
-  ReactDOM.unstable_batchedUpdates(fn);
-}
